requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pytesseract==0.3.9
requests==2.28.0
Unidecode==1.3.4
urllib3==1.26.9
